Split off logging functions/types
